+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 ## Field Tokens v2.0.0-rc3
 
 - Add CHANGELOG.md
+- Add default theme to tests